[#13] Add second screen for Account Details
