+++ Habitica/res/xml/preferences_fragment.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8"?>
+<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android"
+    android:title="@string/PS_contact_title" >
+
+    <EditTextPreference
+        android:key="accountDetails"
+        android:title="AccountDetails"
+        android:summary="Check your account details" />
+
+
+    <Preference android:title="LOGOUT"
+        android:key="logout"
+        android:summary="Logout your account"/>
+
+</PreferenceScreen>
