--- Habitica/res/xml/preferences_header.xml
@@ -1,6 +1,0 @@
-<preference-headers xmlns:android="http://schemas.android.com/apk/res/android">
-    <header
-        android:fragment="s">
-
-    </header>
-</preference-headers>
